move to src.add examples
